Refactoring and minor movement changes

+++ src/main/java/polina4096/resquake/mixin/MixinEntity.java
@@ -1,7 +1,6 @@
 package polina4096.resquake.mixin;
 
 import net.minecraft.entity.Entity;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.Vec3d;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -14,7 +13,7 @@ public abstract class MixinEntity {
 	@Inject(method = "updateVelocity", at = @At("HEAD"), cancellable = true)
 	public void updateVelocityInject(float speed, Vec3d movementInput, CallbackInfo ci) {
 		Entity player = (Entity)(Object)this;
-		if (ReSquakeClient.INSTANCE.updateVelocity(player, speed, (float)movementInput.x, (float)movementInput.z)) {
+		if (ReSquakeClient.INSTANCE.updateVelocity(player, speed, movementInput.x, movementInput.z)) {
 			ci.cancel();
 		}
 	}
